requirements.txt
numpy~=1.23.1
Pillow~=9.2.0
h5py~=3.7.0
matplotlib~=3.5.2
setuptools~=58.1.0